requirements.txt
streamlit>=1.44.0
openai>=1.75.0
pydantic>=2.11.0
python-dotenv>=1.0.1
jinja2>=3.1.6
